app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "numpy>=2.3.2",
    "onnx>=1.19.0",
    "onnxruntime-gpu[cuda,cudnn]>=1.22.0",
    "opencv-python>=4.11.0.86",
    "python-multipart>=0.0.20", 
    "pyyaml>=6.0.2",
    "tifffile>=2025.8.28",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.5",
    "onnxconverter-common>=1.16.0",
]
